Rename `solve` to `solve-ps`
Now it is on PyPI !

There was a challenge about naming.
Because `solve` is already used, so i have to find another name.

So i decided to use `solve-ps` instead.
While working so, I found some interesting things about naming.

`-` and `_` is equivalent in python package name.
ex) `pip install scikit-learn` == `pip install scikit_learn`

however, hyphen `-` is not allowed as a identifier in python.
so, if package name contains `-`,
  we cannot import anything from that package !

but curiously, it seems to be common to use `-` in package name.
  and it is even encouraged to normalize `[-_.]` to `-` ! (PEP 503)
in contrast, every package name's `-` is converted to `_` at internal directory name.

Summary
1. `-` and `_` is equivalent in python package name.
2. `-` is not allowed as a identifier in python.
   but it is encouraged to use `-` in package name.
3. internally, `-` is converted to `_` in package name.

references:
- https://peps.python.org/pep-0426/#name
- https://peps.python.org/pep-0503/#normalized-names

Issue: #1